requirements.txt
requests>=2.31.0
websocket-client>=1.6.1
pyairtable>=2.3.3
python-dotenv>=1.0.1
